requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
